requirements.txt
pytest>=7.4.0
pytest-xdist==2.5.0
selenium
allure-pytest
Faker
python-dotenv
flake8